+++ popkorn-example/src/main/kotlin/cc/popkorn/example/model/D.kt
@@ -3,6 +3,7 @@ package cc.popkorn.example.model
 import cc.popkorn.annotations.*
 import cc.popkorn.core.Provider
 import cc.popkorn.Scope
+import cc.popkorn.core.Injector
 
 @Injectable(alias = "is1") @ForEnvironments("env1") class D1: DiA, DiG
 
@@ -29,19 +30,19 @@ class D10: Wrapper.DiD
 @InjectableProvider(alias = "is6")
 @ForEnvironments("env1", "env2", "env3")
 class DCustom6 : Provider<D6> {
-    override fun create(environment: String?) = D6()
+    override fun create(injector: Injector, environment: String?) = D6()
     override fun scope() =  Scope.BY_NEW
 }
 
 @InjectableProvider(alias = "is7")
 @ForEnvironments("env4")
 class DCustom7 : Provider<D7> {
-    override fun create(environment: String?) = D7()
+    override fun create(injector: Injector, environment: String?) = D7()
     override fun scope() =  Scope.BY_USE
 }
 
 @InjectableProvider(alias = "is8")
 class DCustom8 : Provider<D8> {
-    override fun create(environment: String?) = D8()
+    override fun create(injector: Injector, environment: String?) = D8()
     override fun scope() =  Scope.BY_APP
 }