+++ popkorn/src/main/kotlin/cc/popkorn/instances/NewInstances.kt
@@ -1,5 +1,6 @@
 package cc.popkorn.instances
 
+import cc.popkorn.core.Injector
 import cc.popkorn.core.Provider
 
 
@@ -10,10 +11,10 @@ import cc.popkorn.core.Provider
  * @author Pau Corbella
  * @since 1.0
  */
-internal class NewInstances<T:Any>(private val provider: Provider<T>): Instances<T> {
+internal class NewInstances<T:Any>(private val injector: Injector, private val provider: Provider<T>): Instances<T> {
 
     override fun get(environment:String?) : T{
-        return provider.create(environment)
+        return provider.create(injector, environment)
     }
 
     override fun size() = 0