+++ popkorn/src/main/kotlin/cc/popkorn/instances/PersistentInstances.kt
@@ -1,5 +1,6 @@
 package cc.popkorn.instances
 
+import cc.popkorn.core.Injector
 import cc.popkorn.core.Provider
 
 
@@ -10,11 +11,11 @@ import cc.popkorn.core.Provider
  * @author Pau Corbella
  * @since 1.0
  */
-internal class PersistentInstances<T:Any>(private val provider: Provider<T>): Instances<T> {
+internal class PersistentInstances<T:Any>(private val injector: Injector, private val provider: Provider<T>): Instances<T> {
     private val instances = HashMap<String?, T>()
 
     override fun get(environment:String?) : T{
-        return instances.getOrPut(environment) { provider.create(environment) }
+        return instances.getOrPut(environment) { provider.create(injector, environment) }
     }
 
     override fun size() = instances.size