+++ popkorn/src/main/kotlin/cc/popkorn/instances/VolatileInstances.kt
@@ -1,5 +1,6 @@
 package cc.popkorn.instances
 
+import cc.popkorn.core.Injector
 import cc.popkorn.core.Provider
 import java.lang.ref.WeakReference
 
@@ -12,11 +13,11 @@ import java.lang.ref.WeakReference
  * @author Pau Corbella
  * @since 1.0
  */
-internal class VolatileInstances<T:Any>(private val provider: Provider<T>): Instances<T> {
+internal class VolatileInstances<T:Any>(private val injector: Injector, private val provider: Provider<T>): Instances<T> {
     private val instances = HashMap<String?, WeakReference<T>>()
     
     override fun get(environment:String?) : T{
-        return instances[environment]?.get() ?: provider.create(environment).also { instances[environment] = WeakReference(it) }
+        return instances[environment]?.get() ?: provider.create(injector, environment).also { instances[environment] = WeakReference(it) }
     }
 
     override fun size() = instances.size