+++ popkorn/src/test/kotlin/cc/popkorn/data/TestClassByUse.kt
@@ -1,11 +1,12 @@
 package cc.popkorn.data
 
 import cc.popkorn.Scope
+import cc.popkorn.core.Injector
 import cc.popkorn.core.Provider
 
 class TestClassByUse(environment: String?) : TestClassNoProvider(environment)
 
 class TestClassByUse_Provider : Provider<TestClassByUse> {
-    override fun create(environment: String?): TestClassByUse = TestClassByUse(environment)
+    override fun create(injector: Injector, environment: String?): TestClassByUse = TestClassByUse(environment)
     override fun scope() = Scope.BY_USE
 }